Discard replica_not_avaliable error code in partition metadata
The cause of this error code is unknow. Usually when replicas
are offline (not available), the metadata response should give
a incomplete `isr` array, and for metadata request version 5 or
higher, it should also give a non-empty arrray for `offline_replicas`
In some unusual cases, kafka may give a `replica_not_avaliable`
error code (per partition).