Honor LogAppendTime when decoding messages
For messages format version 1:

When LogAppendTime is in use and compression is enabled on the wrapper
message, set the timestamp of each inner message to the timestamp of the
wrapper message.

This is documented in KIP-32 [1].

For messages format version 2:

When LogAppendTime is in use, set MaxTimestamp as the timestamp for all
Records in the RecordBatch. FirstTimestamp + TsDelta always form the
CreateTime, which is unwanted in this case. There's no explicit mention
in the Kafka documentation. See the official Java client implementation
[2], [3].

The fetch test now uses additional topic with the LogAppendTime
configuration enabled. The test checks that the log append time returned
when producing a message batch matches the timestamp in the fetched
messages. It also checks that the create timestamp set by the producer
is different from the timestamp in the fetched message. The create
timestamp is lowered by by a random number to guarantee that it's in the
past and that Kafka's log append time is greater. Randomizing the
timestamp also makes the test which checks the create timestamp more
robust.

[1] https://cwiki.apache.org/confluence/display/KAFKA/KIP-32+-+Add+timestamps+to+Kafka+message#KIP-32-AddtimestampstoKafkamessage-AddaTimestampfieldtothemessageformatwithmaximumallowedtimedifferenceconfigurationonbroker.
[2] https://github.com/apache/kafka/blob/1f1179ea64bbaf068d759aae988bd2a6fe966161/clients/src/main/java/org/apache/kafka/common/record/DefaultRecordBatch.java#L558
[3] https://github.com/apache/kafka/blob/1f1179ea64bbaf068d759aae988bd2a6fe966161/clients/src/main/java/org/apache/kafka/common/record/DefaultRecord.java#L330-L331